requirements.txt
# Core dependencies with pinned versions for reproducible builds
python-dotenv>=1.0.0,<2.0.0
playwright>=1.40.0,<2.0.0
gspread>=5.12.0,<7.0.0
google-auth>=2.25.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
xlrd>=2.0.0,<3.0.0
requests>=2.31.0,<3.0.0
